requirements.txt
llama-cloud-services==0.6.0
llama-index-core==0.12.16.post1
pandas==2.2.3
python-dotenv==1.0.1
XlsxWriter==3.2.2

